Relocate baseActivity model package to base package
