+++ 2003/kangraemin/app/src/main/java/com/example/kangraemin/LoginActivity.kt
@@ -2,7 +2,7 @@ package com.example.kangraemin
 
 import android.content.Intent
 import android.os.Bundle
-import com.example.kangraemin.model.KangBaseActivity
+import com.example.kangraemin.base.KangBaseActivity
 import com.jakewharton.rxbinding3.view.clicks
 import com.jakewharton.rxbinding3.view.focusChanges
 import com.jakewharton.rxbinding3.widget.textChanges
