+++ 2003/kangraemin/app/src/main/java/com/example/kangraemin/MainActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.view.View
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.kangraemin.adapter.SearchResultAdapter
-import com.example.kangraemin.model.KangBaseActivity
+import com.example.kangraemin.base.KangBaseActivity
 import com.example.kangraemin.model.MovieSearchInterface
 import com.example.kangraemin.util.NetworkUtil
 import com.example.kangraemin.util.RetrofitClient
