+++ 2003/kangraemin/app/src/main/java/com/example/kangraemin/base/KangBaseActivity.kt
@@ -1,4 +1,4 @@
-package com.example.kangraemin.model
+package com.example.kangraemin.base
 
 import androidx.appcompat.app.AppCompatActivity
 import io.reactivex.disposables.CompositeDisposable
